feat: Extend ATPTag
- Extend sequence and acknowledgment numbers to uint32_t
- Add workerId field to ATPTag
- Add CopyFrom method to ATPTag
- Update serialization methods for new fields

This change provides better support for larger sequence numbers
and improved packet identification.

+++ contrib/atp/model/atp-header.cc
@@ -84,7 +84,7 @@ ATPHeader::GetJobId() const
 }
 
 void
-ATPHeader::SetSeqNumber(uint8_t seqNum)
+ATPHeader::SetSeqNumber(uint32_t seqNum)
 {
     NS_LOG_FUNCTION(this << seqNum);
     m_seqNum = seqNum;
